F11 extra button added + Texture Rework 1/3 + Some unused data removed
I have to remove more data that isn't necessary and rework the other textures. I just need to rework the BF and Byron textures for the dialogue and delete data, then we can upload it. Delete the files in the "assets/exclude". They aren't needed at all. Just dead weight for the source code.